server/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2

# LLM and embeddings
llama-cpp-python>=0.2.0; platform_system != "Darwin"
llama-cpp-python[metal]>=0.2.0; platform_system == "Darwin"  # Metal support for Apple Silicon
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.21.0
coremltools>=7.0; platform_system == "Darwin"  # Apple Neural Engine support
optimum[ane]>=1.16.0; platform_system == "Darwin"  # Apple Neural Engine optimizations

# File processing
python-docx>=0.8.11
PyMuPDF>=1.19.0
PyYAML>=6.0
python-pptx>=0.6.21
python-magic>=0.4.27
markdown>=3.3.7

# Utilities
python-dotenv>=0.19.0
aiofiles>=0.8.0
tqdm>=4.65.0
psutil>=5.8.0
watchdog>=2.1.9
tenacity>=8.0.1

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.17.0
black>=22.3.0
isort>=5.10.0
mypy>=0.910
flake8>=4.0.0

# Optional GPU support
--find-links https://download.pytorch.org/whl/nightly/cpu/torch_nightly.html 